Autosave implemented, DataGrid changes are now reflected in Deck object.
Autosave now works, and will occur when after AddCard is closed, the delete button is clicked (after the card is removed from the Deck), and when the “save edit” button is clicked.

The DataGrid now updates the List variable cards of currentDeck, so the internal Deck variable should be consistently accurate with what’s shown on the DataGrid.

I updated the import function to only look for two “rows” (word & definition) instead of three since it’s only pulling from “Terms.txt”